requirements.txt
python-dotenv == 1.*
openai == 0.27.*